// apps/resume-app/src/lib/types/Resume.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PersonalInfo } from "./PersonalInfo";
import type { ResumeMeta } from "./ResumeMeta";
import type { Section } from "./Section";

export type Resume = { meta: ResumeMeta, personal: PersonalInfo, sections: Array<Section>, };